backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlmodel==0.0.16
pydantic==2.5.3
python-multipart==0.0.6
psycopg2-binary==2.9.9
asyncpg==0.29.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
PyJWT==2.8.0
python-jose==3.3.0
bcrypt==4.1.2
passlib[bcrypt]==1.7.4
openai==1.12.0
mcp-server==1.0.0
cohere==5.5.0
